QuestionsExams/requirements/base.txt
django==5.0.1
django-environ==0.10.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-filter==23.5
django-autoslug==1.9.8
django-countries==7.5.1
django-phonenumber-field==7.0.2
phonenumbers==8.13.7
drf-yasg==1.21.5
Pillow==10.2.0
argon2-cffi==21.3.0
pytz==2022.7.1
django-taggit==3.1.0
redis==4.5.1
celery==5.2.7
flower==1.2.0
django-celery-email==3.0.0
django-elasticsearch-dsl==7.3.0
django-elasticsearch-dsl-drf==0.22.5
django-admincharts==0.4.1
django-import-export==3.3.7
drf-spectacular>=0.22.1,<0.23
django_extensions
social-auth-app-django==5.4.1
django-allauth
pandas
dj-rest-auth
djangorestframework-simplejwt
google-auth
google-auth-oauthlib
google-auth-httplib2